+++ views/authors/newAuthor.ejs
@@ -5,7 +5,7 @@
 		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
 		<title>Inventory | <%= title %></title>
 		<link rel="stylesheet" href="/styles/index.css" />
-		<script type="module" src="/scripts/authorFormHandler.js"></script>
+		<script type="module" src="/scripts/form-handlers/authorFormHandler.js"></script>
 	</head>
 	<body>
 		<%- include("../partials/header", { backAnchor: true }) %>
